Implement pure-JS General Transit Feed Specification Reference serialization.
The final zipping stage has not yet been automated, nor have the output files been tested,
but they appear to be generally correct.

I have treated the microbus and tempo as separate agencies for the purpose of GTFS,
and all of the timing/frequency/calendar information is fabricated.

A few data-quality issues were uncovered:
- route.transport is very inconsistent, and includes "hiking"
- The "ref" node in routes is sometimes "-" or such instead of blank, when non-applicable.

I couldn't tell if route.stops are ordered, but require such an invariant.